+++ app.js
@@ -1,0 +1,51 @@
+const createError = require('http-errors');
+const express = require('express');
+const path = require('path');
+const logger = require('morgan');
+const dotenv = require('dotenv');
+const db = require('./services/database');
+
+dotenv.config();
+
+const indexRouter = require('./routers/indexRoutes');
+const contactRouter = require('./routers/contactRoutes');
+
+const app = express();
+
+app.set('views', path.join(__dirname, 'views'));
+app.set('view engine', 'ejs');
+
+app.use(logger('dev'));
+app.use(express.json());
+app.use(express.urlencoded({ extended: false }));
+
+app.use('/assets', express.static(path.join(__dirname, 'assets')));
+
+app.use('/api/contact', contactRouter);
+
+app.get('/robots.txt', (req, res) => {
+  res.sendFile(path.join(__dirname, 'robots.txt'));
+});
+
+app.get('/sitemap.xml', (req, res) => {
+  res.sendFile(path.join(__dirname, 'sitemap.xml'));
+});
+
+app.use('/', indexRouter);
+
+app.use((req, res, next) => {
+  next(createError(404));
+});
+
+app.use((err, req, res, next) => {
+  res.locals.message = err.message;
+  res.locals.error = req.app.get('env') === 'development' ? err : {};
+  res.status(err.status || 500);
+  res.render('error');
+});
+
+db.init().catch((err) => {
+  console.error('[DB] Initialization failed:', err.message || err);
+});
+
+module.exports = app;